fix: address A2A outbound claude review
